refactor some plotting code
