+++ .entangled/tasks.json
@@ -102,6 +102,63 @@
     {
       "description": null,
       "creates": [
+        "data/alcaps_default.h5"
+      ],
+      "requires": [
+        "src/Model/ALCAPS.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/alcaps/defaults.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "docs/src/_fig/alcaps_default_profile.png"
+      ],
+      "requires": [
+        "ext/VisualizationExt.jl",
+        "data/alcaps_default.h5"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/alcaps/plot-defaults.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "data/alcaps2.h5"
+      ],
+      "requires": [
+        "src/Model/ALCAPS.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/alcaps/alternative.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "docs/src/_fig/alcaps-alternative.png"
+      ],
+      "requires": [
+        "data/alcaps2.h5"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/alcaps/plot-alternative.jl"
+    },
+    {
+      "description": null,
+      "creates": [
         "data/caps-osc.h5"
       ],
       "requires": [
@@ -172,63 +229,6 @@
       "path": "examples/production-only/plot-cap-slope.jl"
     },
     {
-      "description": null,
-      "creates": [
-        "data/alcaps_default.h5"
-      ],
-      "requires": [
-        "src/Model/ALCAPS.jl"
-      ],
-      "runner": "julia",
-      "stdout": null,
-      "stdin": null,
-      "script": null,
-      "path": "examples/alcaps/defaults.jl"
-    },
-    {
-      "description": null,
-      "creates": [
-        "docs/src/_fig/alcaps_default_profile.png"
-      ],
-      "requires": [
-        "ext/VisualizationExt.jl",
-        "data/alcaps_default.h5"
-      ],
-      "runner": "julia",
-      "stdout": null,
-      "stdin": null,
-      "script": null,
-      "path": "examples/alcaps/plot-defaults.jl"
-    },
-    {
-      "description": null,
-      "creates": [
-        "data/alcaps2.h5"
-      ],
-      "requires": [
-        "src/Model/ALCAPS.jl"
-      ],
-      "runner": "julia",
-      "stdout": null,
-      "stdin": null,
-      "script": null,
-      "path": "examples/alcaps/alternative.jl"
-    },
-    {
-      "description": null,
-      "creates": [
-        "docs/src/_fig/alcaps-alternative.png"
-      ],
-      "requires": [
-        "data/alcaps2.h5"
-      ],
-      "runner": "julia",
-      "stdout": null,
-      "stdin": null,
-      "script": null,
-      "path": "examples/alcaps/plot-alternative.jl"
-    },
-    {
       "name": "figures",
       "requires": [
         "docs/src/_fig/bs92-fig8.svg",
@@ -237,10 +237,10 @@
         "docs/src/_fig/eca.png",
         "docs/src/_fig/life.gif",
         "docs/src/_fig/boundary_types.png",
-        "docs/src/_fig/b13-capsosc-crosssection.png",
-        "docs/src/_fig/b13-crosssection.png",
         "docs/src/_fig/alcaps_default_profile.png",
-        "docs/src/_fig/alcaps-alternative.png"
+        "docs/src/_fig/alcaps-alternative.png",
+        "docs/src/_fig/b13-capsosc-crosssection.png",
+        "docs/src/_fig/b13-crosssection.png"
       ]
     }
   ]
