+++ docs/src/getting-started.md
@@ -1,0 +1,6 @@
+# Getting Started
+
+## TODO
+
+- [ ] Xianhi: outline a user story
+- [ ] Johan: implement
