+++ entangled.toml
@@ -1,3 +1,4 @@
 version="2.0"
 watch_list=["docs/src/**/*.md"]
+ignore_list=["docs/**/.*.md"]
 hooks=["quarto_attributes", "brei"]
