+++ ext/AgeDepthModel.jl
@@ -1,0 +1,11 @@
+# ~/~ begin <<docs/src/visualization.md#ext/AgeDepthModel.jl>>[init]
+module AgeDepthModel
+
+using Makie
+using Unitful
+
+using CarboKitten.Visualization
+using CarboKitten.Export
+
+end
+# ~/~ end
