+++ src/Visualization.jl
@@ -13,5 +13,8 @@ wheeler_diagram!(args...) = print_instructions("wheeler_diagram!", args)
 wheeler_diagram(args...) = print_instructions("wheeler_diagram", args)
 production_curve(args...) = print_instructions("production_curve", args)
 production_curve!(args...) = print_instructions("production_curve!", args)
+stratigraphic_column!(args...) = print_instructions("production_curve!", args)
+age_depth_model!(args...) = print_instructions("age_depth_model!", args)
+
 end  # module
 # ~/~ end
